+++ packages/@leight/source/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/source
 
+## 0.5.36
+
+### Patch Changes
+
+- Improved source/cursor stuff
+
 ## 0.5.35
 
 ### Patch Changes